Complete foundations exercise one
I followed the steps to complete the first exercise in the "intro to css" exercise in the foundations section of this repository. In this exercise, I applied CSS using the three methods: external, internal, and inline. I also created a stylesheet via the command line and linked to it from index.html