Complete margin and padding exercise 1
This exercise was another quick and easy one. My main takeaway is that I didn't think to use "auto" for the margin-left value of element three. I manually added pixels until I got the right location, but it's good to know auto can be used in the future. Will need to test this behavior further.