Complete block and inline exercise 2
In this exercise, I practiced more with margin and padding on various elements. I also used the "display" property to adjust the positioning of nested elements. Again, using auto as a margin value was not what I went with at first, but it's becoming more clear when I can/should use it. I'm getting better with the longhand vs. shorthand CSS for margin and padding as well.